app/pyproject.toml
[tool.poetry]
name = "test-itfox"
version = "0.1.0"
description = ""
authors = ["Andrei Rostovtsev <105871945+Sprin94@users.noreply.github.com>"]
readme = "README.md"
packages = [{include = "test_itfox"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2"
djangorestframework = "^3.14.0"
django-cleanup = "^8.0.0"
python-dotenv = "^1.0.0"
black = "^23.3.0"
isort = "^5.12.0"
gunicorn = "^20.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
skip-string-normalization = true
line-length = 90

[tool.isort]
profile = "black"
line_length = 90
